tests: reboot: test flashing flash-one image
Although not updated as often, the flash-one image can also be updated
via the ethtool flashing operation. Extend the current unit test to
include this.

Unfortunately, we can't determine the version number of a binary
flash-one image, so we don't do an explicit version check with that
iteration.

Arguably, we need to extend the driver code to separate this out as
another flash region in the ethtool interface. Right now, however, we
need to test this interface as is since users can actually use it this
way in upstream kernels.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>