tests: swap nffw unload for NSP soft reset
The nffw unload operation is not as fully featured as the NSP reset, for
example, the PCIe monitor is not enabled again after using the nffw
unload operation. This can cause system instability if any PCIe
interrupts are received and neither NFD nor the PCIe monitor is
available to service those.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>