tests: bpf: add xdp performance tests
Add xdp performance tests to measure DUT performance when loaded
with varying sample files.

All tests will be compared against the baseline and if the offload tx rate
drops by more than 2% or the map write rates drops by more than 3% on large
maps, the test will fail as there is a performance regression.

The same tests can be carried out on driver mode but with more lenient pass
limits of 90% pps and 75% map writes for the synthetic tests, this is due to
the cpu having larger variability as other processes are running on the host
at the same time.

Signed-off-by: David Beckett <david.beckett@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>