compat: fix compilation on Ubuntu RC kernels
Dirk reports Ubuntu rc kernels have letters in the version:

UTS_UBUNTU_RELEASE_ABI 041600rc7

which makes it impossible to treat the define as an integer
constant.  Add a special check for such kernels in the makefile.
When encountered treat the kernel as vanilla (which Ubuntu
pre-release kernels should be).

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Tested-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>