tests: bug fix in kernel comparison function
The kernel_ver_ge function should return true if the tested kernel version
is greater than or equal to the input major and minor version. There was
a bug whereby it would return true if the compared version was less than
the input params but the major number was the same.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>