tests: flower: stabilise tunnel action tests
The VXLAN and Geneve action tests use skip_sw rules to ensure hw offload
is being tested. However, this means that at least the first packet to be
encapsulated will be sent to the host (dropped) but trigger a neighbour
request message. This can be lead to some unpredictability where more than
one packet is sent to the host or where all packets are handled in fw due
to a neighbour entry being present from a previous test.

Prime the neighbour entry table by sending packets before starting the
test itself. All test packets will then be handled on the card making the
outcome consitent. Include a filter on the tcpdump captures to ensure we
only capture the expected tunnel packets.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>