nfp: use PluDevice register for model for non-NFP6000 chips
On NFP6000 and variant devices, the PluDevice register provides the
major chip number (0x6200) and revision number, which alone isn't
enough to correctly identify the model.

However, the model number for NFP3800 and newer devices can be
completely derived from this register only without the additional
information in the SOFTMODEL0 register.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>