tools: pull nfp-drv-kmods over git native transport instead of ssh
Using the git:// URL to clone and pull does not involve authentication
or encryption, which is supposed to make it slightly faster [0]; but the
main advantage here is that it avoids typing passwords in the middle of
the tests if no SSH key for accessing hg.netronome.com has been set for
the machine and for the account in use.

Note that the repo accessed from the git:// URL is read-only, which
should not be an issue for pulling and running the tests.

[0] https://git-scm.com/book/en/v2/Git-on-the-Server-The-Protocols

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Reviewed-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>