tests: bpf: move optimisation flag to CLANG_FLAGS Makefile variable
Move the optimisation flag ("-O2") to the CLANG_FLAGS variable, so that
it can be changed from the command line when compiling the programs.

This flag is NOT supposed to be changed when running the tests at this
time.

It simply turns out that it can be useful sometimes to compile the
programs with a different flag, typically "-O0", when hacking the tests,
for example when trying to produce non-optimised eBPF assembly. This
patch makes it possible without having to edit the Makefile each time
the flag is changed.

The -O2 flag is now called by default on the command line used for
compiling eBPF assembly as well, but it is ignored in that case, so
it makes no difference on that side.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>