tests: ebpf: add test with simplest BPF-to-BPF function call
This simple test performs a very basic BPF-to-BPF function call, in
order to check that the driver can successfully compile such calls.

It comes with a new test class based on XDPpassBaseWithCodegenScan, so
that we can check that the compiled eBPF code actually contains
subprograms (identified through the use of RTN instructions).

The file to compile is written in eBPF assembly, but future tests using
the same class may come as C code, so change XDPpassBaseWithCodegenScan
method get_jit_patterns_file_name() to search for both .S and .c files
when scanning the sources for codegen checks directives.

One consequence on the aforementioned change is that we also try to dump
JIT-ed instructions for tests with eBPF programs compiled from C source
in the eBPFtest class. As some of those tests are expected to fail at
load time, there are actually no instructions to be dumped: trying to
dump them all the same makes bpftool return an error, thus making the
tests fail. To prevent that, change the check_bpf_jit_codegen() method
so it only tries to dump the instructions if a pattern to check was
found in the source files (be it .c or .S) for the eBPF program used in
the test.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>