tests: ebpf: add a second test with (less simple) BPF-to-BPF calls
This tests involves a program with three eBPF functions (in addition to
the main function), for a total of four calls. Calls are not nested. The
objective is to check that the JIT-compiler can successfully compile a
program with several BPF-to-BPF calls.

Passing pointers to data to the functions, as well as filling and
passing a buffer to those functions, is used in order to get something
complicated enough for the clang compiler not to optimise out the
function calls when called with -O2 optimisation level.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>