tests: ebpf: add a test for BPF calls with twelve calls, using CSR
This test was initially created to measure the overhead produced for
saving return addresses and callee-saved registers in subprograms. This
is a version in which some of the callees do use R6, R7, R8 and/or R9.
This program was then kept as an additional test with twelve independent
calls ("flat" call graph).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>