tests: ebpf: check limit of stack size, reject due to R6~R9 saving
This program should work for host XDP, where it requires 512 bytes of
stack. It is really similar to the test for stack size limit introduced
in a previous commit. However, in this program one of the callees uses
R6, thus necessitating that the four callee-saved registers be saved
into its stack frame. As a consequence, the stack frame is higher than
32 bytes, and will be rounded at 64 four bytes. This makes a total of
544 bytes required, so we cannot accept it for offload.

Suggested-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>