tests: ebpf: add assembly program with BPF calls and perf_event_output()
This is a test with a program calling the bpf_perf_event_output() helper
from a subprogram, in order to dump the content of R2 to user space. The
objective is to check that helpers and BPF-to-BPF calls can work
together.

It can also be reused as an example of how to use this helper and perf
event arrays in assembly. Or be .include-d in other assembly programs,
and the subprogram be called to dump R2 in another sample.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>