tests: flower: add max entries test to flower test set
Introduce new test to verify the correct installation of the maximum
repeated hash entries.

We do this using TCs batch command. The alternative of calling
'tc filter add' 500K times is not feasible as this takes too long.
Creating the batch file on the orchestrator and copying it over is
also not feasible as this file can become very large.

Signed-off-by: Pieter Jansen van Vuuren <pieter.jansenvanvuuren@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>