tests: ebpf: fix test xdp_function_call_helpers_adj_head
The test is broken on multiple levels. To begin with, it does not call
the correct test class, and therefore it does not load the correct
eBPF program. Since the program that is loaded instead has no issue,
it masked the other errors.

Other errors include checks that are not performed correctly in the
program ("new_len" would be "old_len - 16", not "old_len + 16"),
incorrect codegen check pattern (the "rtn" instruction appears just
once). Even worse, as the two functions are in the same ELF section, the
first one appearing in the file would be the one injected as a main
function (instead of the one called by the main function).

Fix all these issues by creating and using a relevant test class based
on the one used for function calls, move the main function at the top of
the eBPF program, and fix the program itself.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>