tests: abm: update tests after drop support was added
Not setting the ECN flags was used throughout the tests
as a way of breaking offload.  Now that its a valid
configuration we need to find different ways.

We will also switch to the MARK or DROP action in the FW
which corresponds to what kernel would have done.  Old
FWs passed packets which were not ECN-capable.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>