tools: build 32-bit kernel with ARCH=i386
When running kmod_test_patch.sh, I faced an issue with the 32-bit build
for net-next. Even after running "make ... defconfig" and "make ...
local_defconfig", running the main "make" command to compile the kernel
would keep asking config questions. In particular, it would ask about
option CONFIG_64BIT which defaults to "yes" and asks subsequent
questions related to the arch, which is somewhat concerning for a 32-bit
kernel - and painful for running the tests (which should require no
intervention after they are launched).

I'm not entirely familiar with the intrisics of the kernel Makefile, but
passing "ARCH=i386", as was done already for the "make *config" steps,
was apparently enough to fix it.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>