tools: print major steps for kmod_test_patch.sh output
The output of kmod_test_patch.sh is hard to read, and in particular it
is not always easy to find what kernel is compiling or, more important,
which kernel version threw an error.

To improve the situation, print in bold yellow the names of the kernel
being compiled, or being used for compiling the module. When information
were already printed, bold-yellow-ify them to make it easier to
distinguish them in the output.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>