tools: add a cleanup function to kmod_test_patch.sh to use on exit
The "cleanup" function used so far does not really cleans up things but
rather sends a notification to the user. There are things we may want to
clean up though: unsetting the "kernels" variable, possibly closing a
file descriptor in the future.

Use the "cleanup" function to actually clean things, call it on exit on
success, and create a separate function for the exit trap in case of
failure.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>