tools: silence compilation output by default for kmod_test_patch.sh
Now that kmod_test_patch.sh indicates its progression by printing its
major steps, we can silence most of the output - everything, in fact,
that results from compiling kernels and modules and that is not an error
or a warning.

To do so, pass the "-s" flag to most of the "make" commands when
building the kernel. This does not silence warnings and errors, but most
of the stuff that is not needed. My opinion is that it makes output
much, much (much) more readable, so enable it by default.

When building the modules, things are a bit more tricky because we want
to keep logging even standard output to build.log. Therefore we use some
bash magic to open a file descriptor pointing to build.log, and perform
several redirections:

- stderr is first redirected to stdout, and then appended to the file by
  "tee -a".
- stderr is also printed to the console by tee (redirected to stderr).
- stdout is only redirected to the log file, but after its file
  descriptor has been duplicated by stderr (meaning that stderr is not
  redirected to the log file in the first place, only appended by tee).

Of course, allow users to disable the feature by passing a "-v" switch
on the command line.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>