tests: ebpf: move codegen checks to a separate class
Move the checks for code generated by the JIT-compiler to their own
separate class. The initialisation of an instance takes a "dut"
argument, which is the machine to which the bpftool command must be sent
to dump the code to check.

Leave the methods used to get the names of the sources to parse in the
test classes, as all classes do not get the same files. Those methods
will be moved in a later patch. For now, pass the file name as an
argument to the method used to run the checks.

Since this method is now called from a class called JitCodegenCheck,
rename it as simply "check()" (to avoid using things like
"jit_codegen.check_bpf_jit_codegen()")

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jiong Wang <jiong.wang@netronome.com>