tests: ebpf: remove duplicate collection for codegen check patterns
In commit 1fbc86ec7a02 ("tests: ebpf: add test with simplest BPF-to-BPF
function call"), we added patter checks at the beginning of
check_bpf_jit_codegen() method in order to return early if there is no
pattern to search, so that test does not fail when the eBPF program is
expected to be rejected.

But it seems the method now collects twice the same codegen check
patterns from the source file, which is not necessary. Remove second
collection.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Reviewed-by: Jiong Wang <jiong.wang@netronome.com>