tests: ebpf: add a log section for codegen checks
Add a new subsection in log hierarchy for checking steps related to
pattern detection within the instructions generated by the JIT. Also,
print a message in the logs if codegen scan succeeded and has returned
no error. Use a finally: guard to make sure that the section is ended
even if the instructions for codegen checks fail.

This is in order to help checking that codegen checks effectively
occured, and to find more easily the related commands in the detailed
log file.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jiong Wang <jiong.wang@netronome.com>