tools: always remake config in kmod_test_patch.sh for 32bit net-next
Applying the default and local config for 32bit net-next is only done
the first time tests are launched, and is skipped after that if the
net-next-32bit build directory exits.

However, new config options may occasionally need to be set. To avoid
querying input from the user, always remake the configuration steps,
thus aligning the process on what is done for other kernels.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>