tests: ebpf: improve collection of program id for codegen checks
Retrieving the program id prior to running checks on the instructions
generated by the JIT compiler would assume that the last BPF program
loaded on the system is the one we want. This does not seem to be
robust.

Incidentally, the lack of double quotes around the grep argument would
also break tests for those people who use Zsh as the main shell on the
DUT (Zsh expands the caret as a logical NOT, adding double quotes
prevents expansion and passes the caret to grep instead).

Improve the situation by using "tc filter show" or "ip link show" for
retrieving the id of the program.

We have already have a helper for ip that uses JSON to collect XDP
program ids in a clean way; however, JSON with tc seems broken at this
time, so we parse the regular plain output in a new helper in DrvSystem.

Suggested-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>