compat: fix reporting when no program attached
Tobias reports that disabling xdpoffload on old kernels makes
the driver think something is loaded:

$ ip link set dev eth0 xdpoffload off
$ ip link show dev eth0
9: eth0: <BROADCAST,MULTICAST,UP> mtu 1500 xdpoffload qdisc mq state UP mode DEFAULT group default qlen 1000
    link/ether 00:15:4d:13:23:18 brd ff:ff:ff:ff:ff:ff
    prog/xdp
$ ip link set dev eth0 xdpoffload obj pass.o
RTNETLINK answers: Device or resource busy

This is due to a bug in compat code, we assumed that whenever
hw offload flag is set something is loaded.  Flags are updated
on program removal as well, so that's incorrect.

Reported-by: Tobias Karlsson <tobias.karlsson@netronome.com>
Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Quentin Monnet <quentin.monnet@netronome.com>
Reviewed-by: Simon Horman <simon.horman@netronome.com>