tests: flower: check tunnel decap on shared MACs
A MAC address is not a unique identifier for a netdev. Drivers such as
Linux bonds may share one address across multiple netdevs. If a tunnel
packet is destined for that MAC address then it should be decapsulated no
matter which port it hits (as long as the MAC is offloaded). Similarly, if
the destination MAC is not shared then it should only be decapped on the
port to which that MAC is assigned.

Add a test to check that tunnelled packets are decapsulated if the
destination MAC address is shared across multiple ports. Verify that
decap will not take place on an incorrect port if the MAC is not shared.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>