nfp: bump DMA address of free list entry to 48 bits
The top bits of the DMA address in the free list descriptor for older
HW has always been reserved.  The address may be larger for new HW,
so include the reserved bits into the address field.  For old HW the
DMA address will still be small, and therefore the reserved bits will
get set to 0.

Split the helper macros into 40 and 48 bit versions.

This allows us to use the same free list structures for all data path
code.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>