tests: flower: move from NFPKmodGrp to NFPKmodAppGrp class
Some (or maybe most) systems suffer from a BSP bug that results in
link state issues. This typically leads to several flower test
failures. To mitigate these failures we move from inheriting from the
NFPKmodGrp class to the NFPKmodAppGrp class. The difference being that
the driver will only reload at the start of the group of tests instead
of on each test individually. Reloading less often reduces the
occurrence of the BSP bug but does not eliminate the chances of it
happening at the start of the tests.

This patch has the additional advantage of reducing the run time of
the tests.

Signed-off-by: Pieter Jansen van Vuuren <pieter.jansenvanvuuren@netronome.com>