tests: flower: add test for VXLAN encap with IP on internal port
Add a test that checks for VXLAN encapsulation when the tunnel end point's
IP address is moved off the nfp repr and added to an internal port. This
replicates Open vSwitch configurations where the tunnel end point is on
the OvS bridge.

The test adds 2 flower rules. One matches incoming packets, encaps them,
and sends them to a VXLAN port. The second is an egress filter that
matches packets coming from the internal port and sends them out a
physical port. Because the next hop for the VXLAN port is the internal
port, this forces packets to recirculate on the card and trigger a merge
rule.

The 'send to VXLAN' rule is added using the skip_sw flag. This ensures
that the datapath in on the network card. The second rule cannot take
this flag as it relies on indirect block offloads.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>