tests: split C samples from BPF samples
C samples unrelated to BPF are now available for test usage. Split the
copy function out so this can be invoked seperately.

The KTLS samples are needed on both the DUT and the endpoint, so move
this up to the super class. To make this work, also need to move the
tmpdir to the super class.

While at it, fix an initialization bug when declaring the host_a object.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>