tests: flower: add tunnel decap with VLAN test
Add a test to check for tunnel (VXLAN) decap when the end-point IP address
is on an internal port (e.g. OvS bridge). This requires that a
'pre-tunnel' rule is hit along with the tunnel based rule.

Continue the tests to include VLANs outside of the VXLAN packets and
ensure that these are properly accepted/rejected.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>