tools: add an option to apply patches to net-next and skip checks
In order to help testing the driver patches, add an option to the test
script to tell it that only a subset of patches passed on the command
line are for the driver: other patches are for the kernel repository
(net-next).

When this option is passed, and before applying each patch, the script
looks at the patch "Subject:" line and searches for the "kmods" string
in it.

  - If "kmods" is in the subject, then the patch is applied to the
    driver code, and checks are performed as usual.
  - If "kmods" is not found, then the patch is applied to net-next
    instead. Then we skip all checks for that patch.

For example, imagine a patch series touching both kernel and driver, as
follows:

  - 1-kernel.patch, 2-kernel.patch are kernel code
  - 3-nfp.patch is driver code
  - 4-kernel.patch is kernel code
  - 5-kernel-nfp.patch is a mix of kernel and driver code

Let's say that 3-nfp_bis.patch and 5-nfp_bis.patch are the backported
patches with just the driver code. We can test all the patches at once
with this command:

        $ ./tools/kmod_test_patch.sh -K \
                1-kernel.patch \
                2-kernel.patch \
                3-nfp.patch 3-nfp_bis.patch \
                4-kernel.patch \
                5-kernel-nfp.patch 5-nfp_bis.patch

(Patch 3-nfp.patch may be omitted, if 5-kernel-nfp.patch can compile
without it).

The driver code being compiled is always the code in nfp-drv-kmods, so
applying driver patches to the net-next repo does not have any effect
(other than ensuring dependencies are met for later patches).

v2:
- Do not create a specific branch for commits applied to net-next.
- Rebuild net-next after applying patches (for symbols check).
- Update script documentation.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>