tests: flower: remove unnecessary cleanup code in merge tests
The merge tests (with and without VLAN) add an internal port to the DUT
and statically give it an ARP entry. On cleanup, this ARP is removed.

However, the ARP entry removal code is incorrect and applies the delete
to a repr rather than the internal port. Anyway, as the internal port is
removed straight after, the static ARP will be removed automatically.

This code is therefore not needed so just remove it.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>