tests: reload: extend FW loading test
With recent additions to the driver, firmware loading policies can now
be configured.

Verify the correct loading of flash vs disk firmware by utilizing the
dummy NFD sample firmware as the flash firmware.

Usually, users would configure this with devlink regardless of whether
this is the OOT or upstream driver, however, to reduce the dependency on
specific upstream devlink features, we test the HWinfo knobs directly.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>