tests: fix bpf_refcnt by waiting a bit before reading program/maps
The bpf_refcnt tests adds and deletes qdisc/bpf objects, and checks
that the counter of program and maps are the same before and
after. Seems that the delete operation takes longer now, making the
bpftool -p read of programs or map lists to read old information. Fix
this by adding a delay and read a few times before giving up.

Signed-off-by: Pablo Cascón <pablo.cascon@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>