tests: flower: add tunnel merge and VLAN action tests for Geneve and GRE
New helper code has been added to impliment tunnel action tests that use
'merge' and 'in vlan' NFP functionality. Use this to add such tests for
GRE and Geneve.

INTERNAL NOTE:
There is a problem with GRE in VLAN so the test will not be added until
this is fixed.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>