nfp: test_harness: do not check return value of debugfs_create_u8()
A recent patch accepted to Linux 5.5 removes the return value from the
function debugfs_create_u8() and so leads to a compilation error for the
nfp kmods test_harness:

commit 9655ac4aca20 ("debugfs: remove return value of debugfs_create_u8()")

It is argued in other patches but Greg Kroah-Hartman that:
'When calling debugfs functions, there is no need to ever check the
return value.  The function can work or not, but the code logic should
never do something different based on this'.

Remove the check on this return value in the nfp test harness.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>