nfp: test_harness: do not check return value of debugfs_create_u32()
A recent patch accepted for inclusion in Linux 5.6 removes the return value
from the function debugfs_create_u32() and leads to a compilation error for
the nfp kmods test_harness:

2b07021a940c ("debugfs: remove return value of debugfs_create_u32()")

It is argued in other patches but Greg Kroah-Hartman that:
'When calling debugfs functions, there is no need to ever check the
return value.  The function can work or not, but the code logic should
never do something different based on this'.

Remove the check on this return value in the nfp test harness.

Signed-off-by: Simon Horman <simon.horman@netronome.com>
Signed-off-by: Louis Peens <louis.peens@netronome.com>