tests:samples: add KERNEL_HEADERS include path
Currently trying to make in this directory is quite likely to complain
about TLS_1_3_VERSION, as this is quite new and not installed yet in
many /usr/include... directories. Add a KERNEL_HEADERS path that can be
set, this should point to the uapi headers for the kernel.

These can be compiled with make headers_install if you are using a
custom built kernel.

Signed-off-by: Louis Peens <louis.peens@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>