tests: system: fix ethtool -g parsing
The output from "ethtool -g" does not always return a digit value for
"RX Mini" and "RX Jumbo" and in some cases returns the string "n/a".
This can be seen by looking at the show_u32() function used to print each
attribute value in ethtool/netlink/rings.c

This is problematic as the regex search in tests/linux_system.py is only
looking for a match containing digits. This causes tests.netdev to fail.

The fix updates the regex to accommodate for the possible "n/a" return value
and updates the test return casts accordingly.

Signed-off-by: Daniel de Villiers <daniel.devilliers@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>