test: netdev: fix split mode detection
As suggested by Louis Peens, the tests.netdev.devlink_port_show test
fails after upgrading to Kernel v5.11.11. To detect if the card is in
breakout mode, the presence of the word "split" is used in the output
string. However, iproute2 added a feature which causes "split" to match,
even when the card is not in breakout mode.

The fix is to update the detection logic to look for "split_group" instead.

Signed-off by: Daniel de Villiers <daniel.devilliers@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>