workflow: deprecation of ubuntu-16.04 environment in GitHub Actions
The build.yaml workflow made use of the ubuntu-16.04 environment
to run builds against Ubuntu 16.04 kernels before, but GitHub
have announced that the ubuntu-16.04 environment will be removed
in September.

In order to prevent breakage of our workflow, deprecate/remove
support for Ubuntu 16.04 from the driver. Using CentOS 7.5.1804
containers to replace it for the corresponding kernel version
(from 3.8 to 4.10) in "build upstream" and deleting 16.04 test
in "build ubuntu".

Signed-off-by: Yu Xiao <yu.xiao@corigine.com>
Signed-off-by: Simon Horman <simon.horman@corigine.com>