tests: README: add documentation regarding nti-based test env
There is currently no documentation in the tests directory, or
information regarding the nti-based driver tests.

This commit adds some guidelines on getting a working test environment
up and running, as well as running the nti-based tests.

This commit provides the majority of the setup, but depending on the
required test group, additional config and tweaking might be required.
This will need to be covered in future commits.

Signed-off-by: Daniel de Villiers <daniel.devilliers@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>