tests: flower: disable NetworkManager on interfaces
NetworkManager interferes with some tests - in this case
it was upping the PF interface, leading to packets for
some "miss" tests cases reaching the host, which triggered
ICMP replies. This caused some unexpected packets to reach
the endpoint, causing the test to fail. This specific
case was only related to the PF and UDP based tunnels,
but it is possible that NetworkManager can interfere
in other ways in other tests, so disable it on the
other DUT interfaces as well.

A new helper is added to determine the PF name, this
is used in the new check. A previous case checking
for flower on 'host_a' is also updated to use this
helper instead of hardcoded guessing using the
interface name.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>