tests: flower: fix vxlan_whitelist ipv6 failure
ipv6 tunneling was added to both hostcode and firmware
but the whitelist test does not seem to have been updated
for this. Add a bit of infrastructure to get and parse
the flower firmware versions from ethtool, and add
helper functions to compare these versions. Use these
helpers to confirm whether or not the test should expect
ipv6 tunnel configs to be 'in_hw" in "not_in_hw".

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>