tests: mefw: fix toolchain version check
A check has been added previously to make use of a compatibility
version of nfld when linking with the mip_v2 flag. However this
checked for a specific major.minor version of the toolchain, breaking
the FW build again when using a newer toolchain. Update the check
to '>=' for the MINOR version. Don't do this for the MAJOR version
yet, as the expectation is that an update to that would likely
need more invasive changes.

While at it also squash this warning from the linker:
    WARNING: Option '-name' is deprecated. Use -fw_name instead.
by changing to -fw_name as suggested.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Tested-by: Danie du Toit <danie.dutoit@corigine.com>
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>