tests: netdev: add addaptive coalesce pf test
workflow: The effectiveness of coalesce function is verified by
comparing three test results. First, disable coalesce and set
rx-usrs/rx-frames equal to 50/64, this is advantageous to throughput.
Second, disable coalesce and set rx-usrs/rx-frames equal to 1/0,
this is advantageous to latency. Last, enable adaptive coalesce and compare
this result with the previous two, expect the throughput of the last
time is close to the first and the latency is close to the second.

The comparison specifically checks that there is a significant
improvement in latency (50%) compared to the default settings,
and throughput within 90% of the default. This 90% limit is
to allow for some test variance, and still confirms that there
is a significant improvement in latency without negatively
impacting throughput.

Signed-off-by: Shihong Wang <shihong.wang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>