tests: system: fix Ethtool setting conditional check
A previous patch (ca61be4) applied, to fix the parsing for "ethtool -g"
aims to check for the "None" type when applying Ethtool ring settings.
The conditional checks used to do so however do not work as intended
and will be True for the undesired "None" type, leading to the failure
of certain driver tests.

This patch fixes this issue by removing the type() function from the
conditional checks.

Signed-off-by: Daniel de Villiers <daniel.devilliers@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>