tests: netdev: fix incorrect ethtool paramter in coalesce_vf test
The coalesce_vf test makes use of the "ethtool --show-coalesce" command to
check if the coalesce adaptive RX/TX setting is sucessfully enabled. A single
argument  (the vport) is required, but the current implementation provides
both the vport and the ns. This leads to a failure of the coalesce_vf test.

This patch updates the code to only pass the vport argument, as well as cleanup
some commented-out code.

Signed-off-by: Daniel de Villiers <daniel.devilliers@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>