tests: ebpf: update netdevsim creation
The current method used to create the netdevsim device is no longer supported.
It is suggested that 'echo "[ID] [PORT_COUNT]" > /sys/bus/netdevsim/new_device'
should be used instead. This is introduced in commit e05b2d1 of the Linux
kernel.

This patch creates the netdevsim device using this newer method.

Signed-off-by: Daniel de Villiers <daniel.devilliers@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>