nfdk: Set the L3 checksum offloading flag
We found that when doing iperf testing with tcp traffic,
some of the packets have invalid ipv4 checksum, and discarded
by ip stack, it lead to a bad throughput performence. It turn
out that, in the implementation of nfp_nfdk_tx_csum(), only
L4_CSUM is set, we should also set L3_CSUM flag to enable
recaculating of L3 header checksum.

Signed-off-by: Xingfeng Hu <xingfeng.hu@corigine.com>
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>