tests: rand: netconsole fix test
- stop_netperfs:
Change max_fail value for kill_pidfile command of src to
match that of the dut. Value was initially too low and as a result the
src was trying to kill processes that were already dead and thus failed.

- execute:
Random test selection was selecting a test based on a random
number incorrectly and thus never running flip_driver.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Daniel de Villiers <daniel.devilliers@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>