tests: ebpf_perf: send tcp traffic to verify loaded xdp
Tests were inspecting the output of `ethtool -S` to determine
if the xdp which was loaded is functioning correctly, but was
not sending any traffic to the device.

get_ethtool_rate() now calls the tcpping() function to send traffic
to the DUT from the EP.
timeout parameter added to the tcpping() function for use in the case
where it is more useful to ping the DUT for a certain period of time
than to send a specific number of packets.
The count parameter of the hping3_cmd() function was moved to the opts
string. This way, when there is a value for count, as is the default,
the function will still operate as originally intended, but when the
value for count is None, the count value will not be included in the
command and the command will instead run until the timeout.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>